// crypto_wallet/client/src/lib/modules/user/types.ts
enum TYPES {
  REG_REQUEST = 'REG_REQUEST',
  REG_SUCCESS = 'REG_SUCCESS',
  REG_FAILURE = 'REG_FAILURE',

  AUTH_REQUEST = 'AUTH_REQUEST',
  AUTH_SUCCESS = 'AUTH_SUCCESS',
  AUTH_FAILURE = 'AUTH_FAILURE',

  USER_INFO_REQUEST = 'USER_INFO_REQUEST',
  USER_INFO_SUCCESS = 'USER_INFO_SUCCESS',
  USER_INFO_FAILURE = 'USER_INFO_FAILURE'
}

export default TYPES